+++ app/src/main/java/cherryjam/narfu/arkhdialect/adapter/PhotoAttachmentAdapter.kt
@@ -10,6 +10,7 @@ import cherryjam.narfu.arkhdialect.R
 import cherryjam.narfu.arkhdialect.data.AppDatabase
 import cherryjam.narfu.arkhdialect.data.entity.PhotoAttachment
 import cherryjam.narfu.arkhdialect.databinding.ItemPhotoAttachmentBinding
+import cherryjam.narfu.arkhdialect.utils.AlertDialogHelper
 
 class PhotoAttachmentAdapter :
     SelectableAdapter<PhotoAttachmentAdapter.PhotoAttachmentViewHolder>() {
@@ -45,9 +46,18 @@ class PhotoAttachmentAdapter :
                                 true
                             }
                             R.id.delete -> {
-                                Thread {
-                                    AppDatabase.getInstance(context).photoAttachmentDao().delete(attachment)
-                                }.start()
+                                AlertDialogHelper.showAlertDialog(
+                                    context,
+                                    title = context.getString(R.string.delete_photo_title),
+                                    message = context.getString(R.string.delete_photo_message),
+                                    positiveText = context.getString(R.string.delete),
+                                    negativeText = context.getString(R.string.cancel),
+                                    positiveCallback = {
+                                        Thread {
+                                            AppDatabase.getInstance(context).photoAttachmentDao().delete(attachment)
+                                        }.start()
+                                    },
+                                )
                                 true//
                             }
                             else -> false
